Reorganise server code
