+++ server/dataLoader.js
@@ -1,0 +1,5 @@
+const { matchPathname } = require('./importer');
+
+exports.loadData = function(requestUrl) {
+  return Promise.resolve({});
+};
